chore: Restructure program
