+++ src/Data/EmojiList.ts
