+++ src/Data/HdlReplyData.ts
