+++ src/Data/ListEmoji.ts
